[image-optimizer] Extract lightweight transform runtime
